acls: reflect the 4th field in the POSIX.1e ACL entry
Some archivers, if the uname/gname exists on the system during
archiving, use also the 4th field of the ACL entry for archiving
the UID/GID.  Instead of ignoring this field during extraction,
honor it if the --numeric-owner option is specified.

To avoid unnecessary errors we also honor the 4th field if the
user/group doesn't exist on the system by default.

* src/xattrs.c (skip_to_ext_fields): Removed.
(fixup_extra_acl_fields): Instead of skipping the 4th field, use
it's value instead of the 2nd field when appropriate.
(xattrs__acls_set): Don't ignore fixup_extra_acl_fields errors.
(acls_one_line): Also reflect the --numeric-owner when printing
the contents of acls field in --list.