Don't set directory mode with --no-overwrite-dir
Since 14d8fc718f0c872274b90991ee634b0cd8e1a6f0, with --no-overwrite-dir,
tar attempted to temporarily change the existing target directory mode
to be able to create files in it, and then restore it back in order to
preserve the directory metadata.

This approach created several problems:
- The directory metadata were not restored accurately. The operation led
  to a change of mtime and ctime, and sometimes to a loss of some mode
  bits (they got masked by the current umask).
- If it was not possible to change the target directory mode, tar
  errored out, even if the directory being extracted was empty and there
  was actually no need to create files in it.

Fix by removing code that makes the existing target directory writable
with --no-overwrite-dir.

This reverts part of commit 14d8fc718f0c872274b90991ee634b0cd8e1a6f0
"Fix the --no-overwrite-dir option", but the fix for the original
problem ("Option "--no-overwrite-dir" does not work for empty
directories") is preserved.

As a result, extracting a directory with files when the directory
already exists and is not writable now fails with --no-overwrite-dir
(just as it did before 14d8fc718f0c872274b90991ee634b0cd8e1a6f0).
Ensuring that the target directory is writable is now a responsibility
of the user.

See the thread "no-overwrite-dir does not preserve metadata of existing
directories"
(https://lists.gnu.org/archive/html/bug-tar/2024-11/msg00006.html)
for the original bug report, discussion and details.

* src/extract.c (extract_dir): Remove setting and restoring of directory
permissions in --no-overwrite-dir mode.
* tests/extrac23.at: Remove the test for the now-unsupported case, add a
a test case for the problem with loss of mode bits.